// frontend/generated/models/OrderDetailDto.ts
export interface OrderDetailDto {
    /**
     * 
     * @type {number}
     * @memberof OrderDetailDto
     */
    orderDetailId?: number;
    /**
     * 
     * @type {number}
     * @memberof OrderDetailDto
     */
    orderId?: number;
    /**
     * 
     * @type {StockKeepingUnitDto}
     * @memberof OrderDetailDto
     */
    sku?: StockKeepingUnitDto;
    /**
     * 
     * @type {number}
     * @memberof OrderDetailDto
     */
    quantity?: number;
    /**
     * 
     * @type {PromotionalCampaignDto}
     * @memberof OrderDetailDto
     */
    promotionalCampaign?: PromotionalCampaignDto;
    /**
     * 
     * @type {number}
     * @memberof OrderDetailDto
     */
    unitPrice?: number;
    /**
     * 
     * @type {number}
     * @memberof OrderDetailDto
     */
    subTotal?: number;
    /**
     * 
     * @type {number}
     * @memberof OrderDetailDto
     */
    finalTotal?: number;
    /**
     * 
     * @type {SlotDto}
     * @memberof OrderDetailDto
     */
    slot?: SlotDto;
    /**
     * 
     * @type {Date}
     * @memberof OrderDetailDto
     */
    createdAt?: Date;
    /**
     * 
     * @type {Date}
     * @memberof OrderDetailDto
     */
    updatedAt?: Date;
}