// frontend/generated/models/OrderDto.ts
export interface OrderDto {
    /**
     * 
     * @type {number}
     * @memberof OrderDto
     */
    orderId?: number;
    /**
     * 
     * @type {AccountDto}
     * @memberof OrderDto
     */
    account?: AccountDto;
    /**
     * 
     * @type {Array<OrderStatusHistoryDto>}
     * @memberof OrderDto
     */
    orderStatusHistories?: Array<OrderStatusHistoryDto>;
    /**
     * 
     * @type {OrderStatus}
     * @memberof OrderDto
     */
    latestStatus?: OrderStatus;
    /**
     * 
     * @type {Array<OrderDetailDto>}
     * @memberof OrderDto
     */
    orderDetails?: Array<OrderDetailDto>;
    /**
     * 
     * @type {TransactionDto}
     * @memberof OrderDto
     */
    transaction?: TransactionDto;
    /**
     * 
     * @type {VoucherDto}
     * @memberof OrderDto
     */
    voucher?: VoucherDto;
    /**
     * 
     * @type {ShippingInfoDto}
     * @memberof OrderDto
     */
    shippingInfo?: ShippingInfoDto;
    /**
     * 
     * @type {Date}
     * @memberof OrderDto
     */
    createdAt?: Date;
    /**
     * 
     * @type {Date}
     * @memberof OrderDto
     */
    updatedAt?: Date;
    /**
     * 
     * @type {number}
     * @memberof OrderDto
     */
    subTotal?: number;
    /**
     * 
     * @type {number}
     * @memberof OrderDto
     */
    finalTotal?: number;
}